Restructure Slack code to avoid Temporal context issue
